api/requirements.txt
fastapi
uvicorn
pandas
pyarrow
boto3
s3fs
python-dotenv
pyspark
numpy
